test(scripts): split repo_map_html tests into their own file
scripts/test_repo_map.py covered two modules (repo_map.py and
repo_map_html.py) and was heading toward 600 lines once Tasks 7/8 add
renderer tests. Split it: TestBuildPayload, TestRender, and their
make_report fixture move to the new scripts/test_repo_map_html.py.
make_entry is used by tests on both sides of the split, so it moves to
a shared scripts/repo_map_test_helpers.py instead of being duplicated.

Same 74 tests pass in the repo_map* pair before and after (147 across
all of scripts/) - the move loses nothing.
